# flowcatalyst-typescript/apps/platform-frontend/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    ConflictResponse:
      description: Conflict error response (409)
      type: object
      properties:
        code:
          type: string
          description: Error code
          examples:
            - CONFLICT
        message:
          type: string
          description: Human-readable error message
          examples:
            - Resource already exists
        details:
          type: object
          additionalProperties: {}
          description: Additional conflict details
...
    CreateRoleRequest1:
      type: object
      properties:
        applicationCode:
          type: string
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionInputDto1'
        clientManaged:
          type: boolean
...
    DeleteResponse:
      description: Delete operation response
      type: object
      properties:
        id:
          type: string
          description: ID of the deleted resource
          examples:
            - app_0ABC123DEF456
        resourceType:
          type: string
          description: Type of resource deleted
          examples:
            - application
        message:
          type: string
          description: Human-readable message
          examples:
            - Application deleted successfully
...
    ErrorResponse:
      description: Error response for client errors (400)
      type: object
      properties:
        code:
          type: string
          description: Error code for programmatic handling
          examples:
            - VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          examples:
            - Invalid input provided
        details:
          type: object
          additionalProperties: {}
          description: Additional error details
...
    ForbiddenResponse:
      description: Forbidden error response (403)
      type: object
      properties:
        code:
          type: string
          description: Error code
          examples:
            - FORBIDDEN
        message:
          type: string
          description: Human-readable error message
          examples:
            - Insufficient permissions
        requiredPermission:
          type: string
          description: Required permission if applicable
          examples:
            - application:delete
...
    Instant:
      type: string
      format: date-time
      examples:
        - 2022-03-10T16:15:50Z
...
    NotFoundResponse:
      description: Not found error response (404)
      type: object
      properties:
        code:
          type: string
          description: Error code
          examples:
            - NOT_FOUND
        message:
          type: string
          description: Human-readable error message
          examples:
            - Application not found
        resourceType:
          type: string
          description: Type of resource that was not found
          examples:
            - application
        resourceId:
          type: string
          description: ID that was searched for
          examples:
            - app_0ABC123DEF456
...
    PermissionDto:
      type: object
      properties:
        permission:
          type: string
        application:
          type: string
        context:
          type: string
        aggregate:
          type: string
        action:
          type: string
        description:
          type: string
...
    PermissionInputDto1:
      type: object
      properties:
        application:
          type: string
        context:
          type: string
        aggregate:
          type: string
        action:
          type: string
...
    PermissionListResponse:
      type: object
      properties:
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionDto'
        total:
          type: integer
          format: int32
...
    RoleDto1:
      type: object
      properties:
        name:
          type: string
        applicationCode:
          type: string
        displayName:
          type: string
        shortName:
          type: string
        description:
          type: string
        permissions:
          type: array
          uniqueItems: true
          items:
            type: string
        source:
          type: string
        clientManaged:
          type: boolean
        createdAt:
          $ref: '#/components/schemas/Instant'
        updatedAt:
          $ref: '#/components/schemas/Instant'
...
    RoleListResponse2:
      type: object
      properties:
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto1'
        total:
          type: integer
          format: int32
...
    UnauthorizedResponse:
      description: Unauthorized error response (401)
      type: object
      properties:
        code:
          type: string
          description: Error code
          examples:
            - UNAUTHORIZED
        message:
          type: string
          description: Human-readable error message
          examples:
            - Authentication required
...
    UpdateRoleRequest1:
      type: object
      properties:
        displayName:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionInputDto1'
        clientManaged:
          type: boolean
...
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /api/admin/roles:
    get:
      summary: List all available roles
      description: Returns all roles from the database. Filter by application code
        or source.
      operationId: listRoles
      tags:
        - BFF - Role Admin
      parameters:
        - name: application
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse2'
        '400':
          description: Invalid source filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
    post:
      summary: Create a new role
      description: Creates a new role with source=DATABASE. Role name is auto-prefixed
        with application code.
      operationId: createRole
      tags:
        - BFF - Role Admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest1'
        required: true
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDto1'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '409':
          description: Role already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
  /api/admin/roles/permissions:
    get:
      summary: List all available permissions
      description: Returns all permissions from code and database.
      operationId: listPermissions
      tags:
        - BFF - Role Admin
      responses:
        '200':
          description: List of permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
  /api/admin/roles/permissions/{permission}:
    get:
      summary: Get permission details
      operationId: getPermission
      tags:
        - BFF - Role Admin
      parameters:
        - name: permission
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permission details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionDto'
        '404':
          description: Permission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
  /api/admin/roles/{roleName}:
    put:
      summary: Update a role
      description: Updates a role. CODE-sourced roles can only have clientManaged
        updated.
      operationId: updateRole
      tags:
        - BFF - Role Admin
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest1'
        required: true
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDto1'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
    get:
      summary: Get role details by name
      operationId: getRole
      tags:
        - BFF - Role Admin
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Role details with permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDto1'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
    delete:
      summary: Delete a role
      description: Deletes a role. Only DATABASE and SDK sourced roles can be deleted.
      operationId: deleteRole
      tags:
        - BFF - Role Admin
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Role deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '400':
          description: Cannot delete CODE-defined role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
...
info:
  title: FlowCatalyst Platform API
  version: 1.0.0-SNAPSHOT
  description: FlowCatalyst Platform REST API
servers:
  - url: http://localhost:8080
    description: Auto generated value
  - url: http://0.0.0.0:8080
    description: Auto generated value